+++ spec/javascripts/module.spec.js
@@ -50,7 +50,7 @@ describe("application modules", function(){
     describe("and the parent module does not exist", function(){
       beforeEach(function(){
         MyApp = new Backbone.Marionette.Application();
-        lastModule = MyApp.module("Parent.Child");
+        lastModule = MyApp.module("Parent.Child.GrandChild");
 
         lastModule.start();
       });
@@ -63,8 +63,44 @@ describe("application modules", function(){
         expect(MyApp.Parent.Child).not.toBeUndefined;
       });
 
+      it("should create the grandchild module on the parent module", function(){
+        expect(MyApp.Parent.Child.GrandChild).not.toBeUndefined;
+      });
+
       it("should return the last sub-module in the list", function(){
-        expect(lastModule).toBe(MyApp.Parent.Child);
+        expect(lastModule).toBe(MyApp.Parent.Child.GrandChild);
+      });
+    });
+
+    describe("When defining a grand-child module with an un-defined child, and starting the parent directly", function(){
+      var handler;
+
+      beforeEach(function(){
+        handler = jasmine.createSpy("initializer");
+
+        MyApp = new Backbone.Marionette.Application();
+
+        MyApp.module("Parent.Child.GrandChild", function(mod){
+          mod.addInitializer(handler);
+        });
+
+        MyApp.module("Parent").start();
+      });
+
+      it("should create the parent module on the application", function(){
+        expect(MyApp.Parent).not.toBeUndefined;
+      });
+
+      it("should create the child module on the parent module", function(){
+        expect(MyApp.Parent.Child).not.toBeUndefined;
+      });
+
+      it("should create the grandchild module on the parent module", function(){
+        expect(MyApp.Parent.Child.GrandChild).not.toBeUndefined;
+      });
+
+      it("should start the grandchild module", function(){
+        expect(handler).toHaveBeenCalled();
       });
     });